requirements.txt
anyio==3.3.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
black==21.11b1
boto3==1.20.5
botocore==1.23.5
cffi==1.15.0
cfgv==3.3.1
click==8.0.3
databases==0.5.3
distlib==0.3.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
ffmpeg-python==0.2.0
filelock==3.3.2
future==0.18.2
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
identify==2.3.5
idna==3.3
jmespath==0.10.0
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
pamqp==2.3.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.15.0
psycopg2==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytube>=12.1.0
PyYAML==6.0
regex==2021.11.10
rsa==4.8
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlalchemy2-stubs==0.0.2a19
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.10.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
